Convert web module into JSX file and translate into TS
