+++ package/tsconfig.json
@@ -1,6 +1,5 @@
 {
   "include": ["./src/**/*"],
-  "exclude": ["./src/RNCSliderNativeComponent.web.ts"],
 
   "compilerOptions": {
     "jsx": "react",
